refactor: dispatch userId and triviaId on score page to redux if not already there
