+++ src/components/PlayGame/Host/ScoreAnswers.jsx
@@ -2,18 +2,27 @@ import React, { useState, useEffect } from "react";
 import { useDispatch, useSelector } from 'react-redux';
 import { useRouter } from 'next/router';
 import Link from 'next/link';
+// import { useAppSelector, useAppDispatch } from "../../../../../lib/hooks";
+import { useAppSelector, useAppDispatch } from "../../../../lib/hooks";
+import { setGame } from "../../../redux/reducers/playGameSlice";
 
 
 export default function ScoreAnswers(props) {
-  const {questions} = props;
-  const [triviaId, setTriviaId] = useState(null);
-  useEffect(() => {
-    if (questions && !triviaId) {
-      setTriviaId(questions[0].triviaId);
-    }
-  }, []);
+  const {questions, session, triviaGame} = props;
   const router = useRouter();
+  const dispatch = useAppDispatch()
+  const triviaId = useAppSelector(state => state.playGame.triviaId)
+  useEffect(() => {
+    if (session && triviaGame && !triviaId) {
+      dispatch(setGame({
+        hostId: Number(triviaGame.hostId), 
+        triviaId: Number(triviaGame.id),
+        userId: Number(session.user.id)
+      }))
+    }
+  }, [])
   const roundNum = router.query.round.split('-')[1];
+  const joinCode = router.query.joinCode;
   const markAsTrue = async (answerData) => {
     const updatedAnswer = await fetch(
       '/api/update/answers',
@@ -54,7 +63,7 @@ export default function ScoreAnswers(props) {
         ) : <div>No questions have been submitted yet</div>}
         </ul>
       </div>
-      <Link href={`/game/${triviaId}/round-${Number(roundNum) + 1}/admin/play`}>
+      <Link href={`/game/${joinCode}/round-${roundNum}/overview`}>
         <button>Next</button>
       </Link>
     </div>
